refactor: re-organize auth folder in sdk
