+++ index.ts
@@ -19,8 +19,8 @@
  */
 
 export { BaseService } from './lib/base_service';
-export { IamTokenManagerV1 } from './iam-token-manager/v1';
-export { Icp4dTokenManagerV1 } from './auth/icp-token-manager';
+export { IamTokenManagerV1 } from './auth/iam-token-manager-v1';
+export { Icp4dTokenManagerV1 } from './auth/icp-token-manager-v1';
 export * from './lib/helper';
 export { default as qs } from './lib/querystring';
 export { default as contentType } from './lib/content-type';
