+++ lib/base_service.ts
@@ -16,8 +16,8 @@
 
 import extend = require('extend');
 import vcapServices = require('vcap_services');
-import { Icp4dTokenManagerV1 } from '../auth/icp-token-manager';
-import { IamTokenManagerV1 } from '../iam-token-manager/v1';
+import { IamTokenManagerV1 } from '../auth/iam-token-manager-v1';
+import { Icp4dTokenManagerV1 } from '../auth/icp-token-manager-v1';
 import { stripTrailingSlash } from './helper';
 import { readCredentialsFile } from './read-credentials-file';
 import { sendRequest } from './requestwrapper';
